+++ src/main/java/space/space/services/factories/Constants.java
@@ -2,5 +2,5 @@ package space.space.services.factories;
 
 public class Constants {
     public static final long INITIAL_CREDIT_AMOUNT = 10000;
-    public static final int INITIAL_INCOME_PER_TRIP = 100;
+    public static final long INITIAL_INCOME_PER_TRIP = 100;
 }
