+++ src/main/java/space/space/services/services/UsersService.java
@@ -3,6 +3,6 @@ package space.space.services.services;
 import org.springframework.security.core.userdetails.UserDetailsService;
 
 public interface UsersService extends UserDetailsService {
-   void spendMoney(String username, double spentPot) throws Exception;
-   void earnMoney(String username, double spentPot) throws Exception;
+   void spendMoney(String username, long spentPot) throws Exception;
+   void earnMoney(String username, long spentPot) throws Exception;
 }
